Fix FAQ card going under items in page
