+++ dapp/src/components/Footer.tsx
@@ -2,7 +2,7 @@ import { Link } from "react-router-dom";
 import polygonSrc from "assets/images/polygon.png";
 import thegraphSrc from "assets/images/thegraph.png";
 import IPFSClientWidget from "./IPFSClientWidget";
-import FAQButton from "./FAQButton";
+import FAQButton from "./FAQCard";
 
 export default function Footer({
   showContentDisclaimer = false,
